select wallet on click
