+++ packages/bitcore-wallet-ui/src/containers/wallet/wallet.tsx
@@ -4,7 +4,6 @@ import { RouteComponentProps } from 'react-router';
 import { WalletBar } from './BalanceCard';
 import { TransactionListCard } from './TransactionContainer';
 import { WalletBottomNav } from './BottomNav';
-import { socket } from '../../contexts/io';
 import { ActionCreators, store } from '../../index';
 import { connect } from 'react-redux';
 import { AppState } from '../../contexts/state';
@@ -17,36 +16,22 @@ interface Props extends RouteComponentProps<{ name: string }> {
 }
 
 class WalletContainer extends Component<Props> {
-  componentDidMount = async () => {
-    const name = this.props.match.params.name;
-    store.dispatch(ActionCreators.setWalletName(name));
-    const wallet = await this.loadWallet(name);
-    await wallet!.register({ baseUrl: 'http://localhost:3000/api' });
-    if (!this.props.wallet) {
-      await store.dispatch(ActionCreators.setWallet(wallet!));
-    }
-    if (this.props.wallet) {
-      await this.handleGetTx(this.props.wallet);
-      await this.handleGetBlock(this.props.wallet);
-      await this.updateWalletInfo(this.props.wallet);
-      await this.fetchAddresses(this.props.wallet);
+  componentDidUpdate = async (prevProps: any) => {
+    const { wallet } = this.props;
+    if (wallet && prevProps.wallet !== wallet) {
+      await this.updateWalletInfo(wallet);
+      await this.fetchAddresses(wallet);
     }
   };
+
   updateWalletInfo = async (wallet: Wallet) => {
     await this.fetchTransactions(wallet);
     await this.updateBalance(wallet);
   };
 
-  handleGetTx = (wallet: Wallet) => {
-    socket.on('tx', () => {
-      this.updateWalletInfo(wallet);
-    });
-  };
-
-  handleGetBlock = (wallet: Wallet) => {
-    socket.on('block', () => {
-      this.updateWalletInfo(wallet);
-    });
+  updateBalance = async (wallet: Wallet) => {
+    const balance = await wallet.getBalance();
+    store.dispatch(ActionCreators.setBalance(balance));
   };
 
   fetchTransactions = async (wallet: Wallet) => {
@@ -81,27 +66,6 @@ class WalletContainer extends Component<Props> {
           store.dispatch(ActionCreators.setAddress(a.address))
         );
       });
-  };
-
-  updateBalance = async (wallet: Wallet) => {
-    const balance = await wallet.getBalance();
-    await store.dispatch(ActionCreators.setBalance(balance));
-  };
-
-  loadWallet = async (name: string) => {
-    let wallet: Wallet | undefined;
-    try {
-      const exists = Wallet.exists({ name });
-      if (!exists) {
-        console.log('Wallet needs to be created');
-      } else {
-        console.log('Wallet exists');
-        wallet = await Wallet.loadWallet({ name });
-      }
-    } catch (err) {
-      console.log(err);
-    }
-    return wallet;
   };
 
   render() {
